% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_extract_info_from_gtf.R
\name{compute_gene_non_redundant_exonic_length}
\alias{compute_gene_non_redundant_exonic_length}
\title{Compute non-redundant exonic length of genes}
\usage{
compute_gene_non_redundant_exonic_length(gtf_df)
}
\arguments{
\item{gtf_df}{A data frame containing the GTF (Gene Transfer Format) data.
This should include columns such as \code{seqid}, \code{start}, \code{end}, \code{strand}. \cr
GTF files can be loaded
with function \code{rtracklayer::readGFF}}
}
\value{
A data.frame with gene_id and exonic_length columns
}
\description{
For each gene, overlapping regions
from its different exons are first collapsed into
unique genomic regions. The 'non-redundant exonic
length' is then computed as the sum of the lengths
of the genomic regions covered by a gene's exonic
sequences.
}
\examples{
 head(gtf_df)
 compute_gene_non_redundant_exonic_length(gtf_df)

}
