% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_extract_info_from_gtf.R
\name{compute_per_transcript_exonic_length}
\alias{compute_per_transcript_exonic_length}
\title{Calculates exonic length of each transcript}
\usage{
compute_per_transcript_exonic_length(gtf_df)
}
\arguments{
\item{gtf_df}{A data.frame resulting from reading a gtf file, e.g by
rtracklayer::readGFF}
}
\value{
a data.frame with transcript_id and length columns
}
\description{
Calculates exonic length of each transcript
}
\examples{
compute_per_transcript_exonic_length(gtf_df)
}
