% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_extract_info_from_gtf.R
\name{exons_per_transcript}
\alias{exons_per_transcript}
\alias{exons_per_gene}
\title{Get the number of exons per transcript or gene}
\usage{
exons_per_transcript(gtf_df, anno_col = "biotype")

exons_per_gene(gtf_df, gene_col = "gene_id", anno_col = "biotype")
}
\arguments{
\item{gtf_df}{A data frame containing the GTF (Gene Transfer Format) data.
This should include columns such as \code{seqid}, \code{start}, \code{end}, \code{strand}. \cr
GTF files can be loaded
with function \code{rtracklayer::readGFF}}

\item{anno_col}{An annotation column ('biotype' by default) you may want to retrieve together with the exon number}

\item{gene_col}{Name of the column to use to summarise at gene level, default: "gene_id"}
}
\value{
A data.frame with transcript_id/gene_id, N_exons and 'anno_col' (if present in gtf)
}
\description{
The functions \code{exons_per_transcript} and \code{exons_per_gene} retrieve
the exon count for each transcript or gene, respectively. At gene level,
the number of unique exons is retrieved, i.e. identical exons shared by different
transcripts are only counted once.
Additionally, an annotation column ('biotype' by default)
may be retrieved for each transcript or gene if present in then input gtf
}
\examples{
exons_per_transcript(gtf_df)
}
