% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_extract_info_from_gtf.R
\name{get_per_gene_merged_exons}
\alias{get_per_gene_merged_exons}
\title{Get per gene merged exons}
\usage{
get_per_gene_merged_exons(gtf_df)
}
\arguments{
\item{gtf_df}{A data.frame resulting from reading a gtf file, e.g by
rtracklayer::readGFF}
}
\value{
A data.frame with all non-redundant exonic regions covered by each gene
}
\description{
For each gene, overlapping regions
from its different exons are collapsed into
unique genomic regions.
}
\examples{
head(gtf_df)
get_per_gene_merged_exons(gtf_df)
}
