% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtf_df}
\alias{gtf_df}
\title{GTF}
\format{
A data.frame with 19 variables and 1901 observations
}
\usage{
gtf_df
}
\description{
Toy example of a mm39 mouse GTF file as imported with rtracklayer::readGFF\cr
NOTE that, when reading another GTF, not all variables
need to be equivalent to the ones in this example.\cr
Important columns are: "seqid", "start", "end", "strand",
"transcript_id", "gene_id"...\cr
This particular GTF example is a subset with 100 genes and 224 transcripts
form a merged transcriptome generated with gffCompare
from StringTie assemblies on 3 LSK total-RNA replicates
}
\keyword{datasets}
