% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_overlapping_class_codes}
\alias{non_overlapping_class_codes}
\title{Non-overlapping classcodes}
\format{
A character vector of length 7: "i" "s" "r" "u" "x" "y" "."
}
\usage{
non_overlapping_class_codes
}
\description{
classcodes used to classify assembled transfrags present in the tracking
files from the GffCompare utility were classified in overlapping
a reference transcript (in same strand) and non-overlapping
}
\keyword{datasets}
