% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overlapping_class_codes}
\alias{overlapping_class_codes}
\title{Overlapping classcodes}
\format{
A character vector of length 9 : "=" "c" "e" "o" "j" "k" "m" "n" "p"
}
\usage{
overlapping_class_codes
}
\description{
classcodes used to classify assembled transfrags present in the tracking
files from the GffCompare utility were classified in overlapping
a reference transcript (in same strand) and non-overlapping
}
\keyword{datasets}
