% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_extract_info_from_gtf.R
\name{transcripts_per_gene}
\alias{transcripts_per_gene}
\title{Get number of transcripts per gene}
\usage{
transcripts_per_gene(gtf_df, gene_col = "gene_id", anno_col = "biotype")
}
\arguments{
\item{gtf_df}{A data.frame resulting from reading a gtf file, e.g by
rtracklayer::readGFF}

\item{gene_col}{Name of the column to use to summarise at gene level, default: "gene_id"}

\item{anno_col}{An annotation column ('biotype' by default) you may want to retrieve together with the exon number}
}
\value{
A data.frame with gene_id, N_transcripts and 'anno_col' (if present in gtf)
}
\description{
Get number of transcripts per gene
}
\examples{
transcripts_per_gene(gtf_df)
}
