% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_extract_info_from_gtf.R
\name{get_per_gene_merged_exons}
\alias{get_per_gene_merged_exons}
\title{Get per gene merged exons}
\usage{
get_per_gene_merged_exons(gtf_df)
}
\arguments{
\item{gtf_df}{A data frame containing the GTF (Gene Transfer Format) data.
This should include columns such as \code{seqid}, \code{start}, \code{end}, \code{strand}. \cr
GTF files can be loaded
with function \code{rtracklayer::readGFF}}
}
\value{
A data.frame with all non-redundant exonic regions covered by each gene
}
\description{
For each gene, overlapping regions
from its different exons are collapsed into
unique genomic regions.
}
\examples{
head(gtf_df)
get_per_gene_merged_exons(gtf_df)
}
